done till speciality

+++ client/src/components/RouteProtection/protectedRoute.jsx
@@ -1,40 +1,30 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { Navigate } from 'react-router-dom';
-import {useSelector,useDispatch} from 'react-redux'
-import axios from '../../axios/axios';
+import { useDispatch } from 'react-redux';
+import { useGetUserDetailsQuery } from '../../redux/features/api/apiSlice';
 import { setUser } from '../../redux/features/userSlice';
 
 export default function ProtectedRoute({ children }) {
-  const dispatch = useDispatch()
-const {user} = useSelector(state=>state.user)
-
-// eslint-disable-next-line react-hooks/exhaustive-deps
-const getUser = async()=>{
-  try{
-    const res = await axios.post('/getUserData',
-    {token:localStorage.getItem('token')},
-    {
-      headers:{
-        Authorization:`Bearer ${localStorage.getItem("token")}`
+  const dispatch = useDispatch();
+  const token = localStorage.getItem('token');
+  const { data, isSuccess } = useGetUserDetailsQuery();
+  if (isSuccess) {
+    dispatch(setUser(data.data));
+  }
+  if (token) {
+    try {
+      if (data) {
+        if (!data.success) {
+          localStorage.clear();
+         return <Navigate to="/login" />;
+        }
       }
-    })
-    if(res.data.success){
-      dispatch(setUser(res.data.data))
-    }else{
-    <Navigate to='/login'/>
-    localStorage.clear()
+    } catch (err) {
+      console.log('error in protected route', err);
     }
-  }catch(error){
-    console.log(error);
   }
-}
-useEffect(()=>{
-  if(!user){
-    getUser()
+  if (localStorage.getItem('token')) {
+    return children;
   }
-},[user,getUser])
-    if (localStorage.getItem('token')) {
-      return children;
-    }
-    return <Navigate to="/login" />;
+  return <Navigate to="/login" />;
 }
